Move varzerovec into toplevel module
