+++ utils/zerovec/src/varzerovec/serde.rs
