+++ utils/zerovec/src/zerovec/mod.rs
@@ -5,8 +5,6 @@
 #[cfg(feature = "serde")]
 mod serde;
 
-pub(crate) mod var;
-
 use crate::ule::*;
 use std::fmt;
